Refactor calced_props and domain_metrics to use metric registry
- Replace inline calculation logic with registry-driven iteration over
  DAILY_METRICS.

- Rename and move domain_metrics() to compute_daily_metrics_for_domain()

- compute_daily_metrics_for_domain() now filters out
  _MODEL_MANAGED_FIELDS to avoid overwriting derived properties.

- Remove unused imports.
