+++ corehq/apps/data_analytics/metric_registry.py
@@ -1,5 +1,6 @@
 import logging
 from dataclasses import dataclass
+from datetime import datetime, timezone
 from functools import cached_property
 from typing import Any, Callable, Literal
 
@@ -79,3 +80,20 @@ def collect_metrics_for_domain(domain_obj, metrics, existing_metrics=None):
                 metric.field_name, domain_obj.name,
             )
     return updates
+
+
+def compute_daily_metrics_for_domain(domain_obj, all_stats):
+    """Compute all daily metrics for a domain as a dict for DomainMetrics.
+
+    Returns dict of {field_name: value} suitable for update_or_create().
+    """
+    from .daily_calcs import _MODEL_MANAGED_FIELDS, DAILY_METRICS
+
+    ctx = DomainContext(domain_obj, all_stats=all_stats)
+    metrics_dict = {}
+    for metric in DAILY_METRICS:
+        if metric.field_name not in _MODEL_MANAGED_FIELDS:
+            metrics_dict[metric.field_name] = metric.calc_fn(ctx)
+    metrics_dict['domain'] = domain_obj.name
+    metrics_dict['last_modified'] = datetime.now(tz=timezone.utc)
+    return metrics_dict
